Node methods disallow self-reference.
The Node methods API has changed slightly, in that most return the node
except where that is not very logical, such as with the NewChild method.
In particular, the AddChild method returns the instance, rather than the
child.  Logic in the Parse function has been modified accordingly.

Moreover, the `SetParent` and `AddChild` explicitly disallow for
self-reference.  In particular, `n.SetParent(n)` and `n.AddChild(n)`
both do not mutate n.

Node.IsValid now checks whether the parent is the self, and if so,
reports invalidity.  The corresponding unit test is added.